Use <dialog> for binder EditOrRun modal
